Url normalizer+puny encoder
